Isolate flow to a queue -- working patch
Right now it creates only 2 queues, generic handling is done by queue 1 and queues to handles flow of choice. An API for communicating flow specs is needed but for now it is hard coded. Destruction path upon error is buggy and needs to be fixed. Destruction path upon ring closure/interface shut down is missing and needs to be implemented.